protonect: added install target and use pkg-config to find dependancis
The install targets for protonect was missing, now the lib and all headers
got installes.

Also make it configurable wether the internal-find algorithm should
be used, or if it is assumed that the depending libraries (libusb/glfw/glew)
are installed in the system. This needs a fix of the include of libusb itself
in the headers.

All changes should be backward compatible.