Create static Protonect.exe on Windows
Static Protonect.exe does not require managing DLLS. As a result,
only libfreenect2static.lib is provided by default. (Only libusb,
OpenCV, TurboJPEG and GLFW are statically linked now. OpenCL not.)

This is because it is impossible to import both dynamic and static
versions of OpenCV at the same time, also building both dll and
lib requires recompiling objects with /MD and /MT.

libfreenect2static.lib is compiled with /MT by default for static
linking with Protonect.exe. If BUILD_SHARED_LIBS is true, it is
compiled with /MD.

Dynamic verion of Protonect.exe and libfreenect2.dll can be built
with -DBUILD_SHARED_LIBS=ON. Manual DLL management is required
to run dynamic Protonect.exe.