Convert to a global static logger
Before this commit, logger pointers get passed around through
inheritance and manually constructed dependency assignment lists.
The manual management is hard to scale with logging calls which
can appear anywhere in the code.

This commit implements a single global static logger for all
Freenect2 contexts. It still can be replaced by different
loggers, but only one at a time.

Now it is the responsibility of each logging point to include
libfreenect2/logging.h, which is not automatically included.