usb: Print correct firmware version number
Blob #3 is the main one in the firmware's 7 blobs, and should
represent version of other blobs, except the bootloader blobs
which is never updated and not to be bothered with about their
versions.

The official SDK uses only blob #3 to report the version. Use it
for the version number here.