registration: Add depth-only methods - undistortDepth() is the equivalent for apply() but without color - getPointXYZ() is the equivalent for getPointXYZRGB() without color
This commit enables to work only with the depth without having to process the color image.
Indeed, the implementation forces you so far to register the color image if you want
to compute any 3D Point associated with a pixel value. This is time consuming and
critical for applications that require to be run in real time.