opencl: Use precomputed sin/cos tables
Instead of computing the sine and cosine for the p0 table and the
phases on the GPU, they are now precomputed once on the CPU.

Details: Replaced sin(a+b) by sin(a)*cos(b)+cos(a)*sin(b), where
sin(a),cos(b),cos(a),sin(b) are stored in a LUT.  Simplyfied
processPixelStage1 code and removed processMeasurementTriple.

Moved one if from decodePixelMeasurement to processPixelStage1.

Removed the first part of `valid && any(...)` because valid has been
checked before.