Add error propagation for processors
The new internal API policy:

Packet processors should report internal errors by setting
good() to false, and pass the last frame to the user with
status set to 1.

Currently CUDA, OpenCL, Tegra, and VAAPI have been added
with the error propagation. CPU, OpenGL, and VT have no
error checking in place so they do not report errors.
TurboJPEG seems to produce non-fatal errors so it also
does not propagate errors.

The user should check the received frame's status
for errors. If there are errors, the user should stop the
device and exit.

When good() is false, the processor->process()
will no longer be called, and if the user continues to
call waitForNewFrame(), it will hang.