Add KDE depth unwrapping algorithms
This implements kernel density estimation based phase unwrapping
procedure. It shows improved depth imaging, especially for large depth and
outdoors scenes. The method was presented on ECCV 2016, see paper for more
information.

http://users.isy.liu.se/cvl/perfo/abstracts/jaremo16.html

The algorithms are added as OpenCL and CUDA processors. OpenCLKde and CudaKde
pipelines are also added as APIs.