opencl: Use 1.0f float to avoid llvm errors
Beignet recommends:

If you use 1.0 in the kernel, LLVM 3.6 will treat it as 1.0f, a
single float, because the project doesn't support double float.
but LLVM 3.8 will treat it as 1.0, a double float, at the last
it may cause error.  So we recommend using 1.0f instead of 1.0
if you don't need double float.